refactor: 구조 변경 및 RaceManager RaceContoller로 이름 변경
