+++ src/main/kotlin/step3/carRacing/main.kt
@@ -1,5 +1,7 @@
 package step3.carRacing
 
+import step3.carRacing.model.CarRacing
+
 fun main() {
     CarRacing().start()
 }
