+++ src/main/kotlin/step3/carRacing/model/CarRacing.kt
@@ -1,5 +1,9 @@
-package step3.carRacing
+package step3.carRacing.model
 
+import step3.carRacing.Action
+import step3.carRacing.DispatchHandler
+import step3.carRacing.Payload
+import step3.carRacing.contoller.RaceController
 import step3.carRacing.view.InputView
 import step3.carRacing.view.OutputView
 
@@ -8,7 +12,7 @@ class CarRacing {
     private val inputView = InputView(dispatchHandler)
 
     init {
-        RaceManager(dispatchHandler)
+        RaceController(dispatchHandler)
         OutputView(dispatchHandler)
     }
