+++ src/test/kotlin/step3/carRacing/contoller/RaceControllerTest.kt
@@ -1,0 +1,5 @@
+package step3.carRacing.contoller
+
+import org.junit.jupiter.api.Assertions.*
+
+internal class RaceControllerTest
